+++ lib/src/system/types/lu.dart
@@ -1,10 +1,11 @@
 import 'package:equations/equations.dart';
 import 'package:equations/src/system/system.dart';
+import 'package:equations/src/system/utils/matrix_utils.dart';
 
 /// Solves a system of linear equations using the 'LU decomposition' method.
 /// The given input matrix, representing the system of linear equations, must
 /// be square.
-final class LUSolver extends SystemSolver {
+final class LUSolver extends SystemSolver with RealMatrixUtils {
   /// {@macro systems_constructor_intro}
   ///
   ///  - [matrix] is the matrix containing the equations;
@@ -21,11 +22,11 @@ final class LUSolver extends SystemSolver {
     // Solving Ly = b
     final L = lu.first.toListOfList();
     final b = knownValues;
-    final y = SystemSolver.forwardSubstitution(L, b);
+    final y = forwardSubstitution(L, b);
 
     // Solving Ux = y
     final U = lu[1].toListOfList();
 
-    return SystemSolver.backSubstitution(U, y);
+    return backSubstitution(U, y);
   }
 }